+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## 1.12.2 (Unreleased)
 ## 1.12.1 (August 26, 2019)
 
 BUG FIXES: